requirements.txt
Django~=6.0
djangorestframework~=3.16.1
openpyxl~=3.1.5
djangorestframework_simplejwt~=5.5.1
django-mptt~=0.18.0
django-filter~=25.2
django-fsm
django-notifications-hq